Created Site & Dumped Database
Site is BLANK.    There were some complications when building this from
teh Install Profile that was provided.    I am ending my efforts to load
dummy data into this site as I've reached 2 hours of work when the
estimate was 1 hour.    I will await Jason's confirmation and a look
through to make sure that I've not done this incorrectly.    I may
rebuild the site from Drupal Core and load the modules and themes in
manually if that is still causing an issue.